Lots of new stuff. But still no clarity
