+++ analysis/src/main/java/org/hps/analysis/alignment/FeeSvtAlignmentAnalyzer.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package org.hps.analysis.alignment;
 
 import hep.aida.IAnalysisFactory;
